chore: rename defineCustomElement.ts so I can fix it
