+++ src/package/index.ts
@@ -1,7 +1,7 @@
 import { initCache } from './polyfill/hmrCache';
 import { overrideCustomElementDefine } from './polyfill/overrideCustomElementDefine';
 export { reflowDOM } from './utils/reflowDOM';
-export { defineCustomElement } from './decorator/defineCustomElement';
+export { defineCustomElement } from './decorator/xdefineCustomElement';
 export { onCustomElementChange } from './polyfill/onCustomElementChange';
 
 export function applyPolyfill() {
